# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * CS Manager game API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
}

export interface NewGameInput {
  org_id: string;
  manager_name: string;
  year?: number;
}

export interface OrgSummary {
  id: string;
  name: string;
  tag: string;
  region: string;
  tier: number;
  rating: number;
  budget: number;
  roster_size: number;
}

export interface GameState {
  has_game: boolean;
  manager_name?: string;
  week: number;
  month: number;
  year: number;
  season_phase: string;
  org_id?: string;
  org_name?: string;
  org_tag?: string;
  region?: string;
  tier?: number;
  budget: number;
  roster_size: number;
  academy_size?: number;
  form: string[];
  /** @nullable */
  world_rank?: number | null;
  chemistry?: number;
  pressure?: number;
  news_count?: number;
  week_label?: string;
}

export interface MatchResult {
  tournament_name: string;
  team_a: string;
  team_b: string;
  score_a: number;
  score_b: number;
  winner: string;
  is_player_match: boolean;
  maps: string[];
  week: number;
}

export interface NewsItem {
  id: string;
  headline: string;
  category: string;
  week: number;
  month?: number;
  year?: number;
  timestamp?: string;
}

export interface AdvanceResult {
  state: GameState;
  match_results: MatchResult[];
  new_news: NewsItem[];
  week_summary: string;
}

export type PlayerCardAttributes = {[key: string]: number};

export interface PlayerCard {
  id: string;
  name: string;
  alias: string;
  age: number;
  nationality: string;
  role: string;
  rating: number;
  hltv_rating: number;
  salary: number;
  contract_years?: number;
  form: string[];
  mental: string;
  is_academy: boolean;
  /** @nullable */
  org_id?: string | null;
  attributes?: PlayerCardAttributes;
}

export interface RosterData {
  roster: PlayerCard[];
  academy: PlayerCard[];
  budget: number;
  chemistry: number;
  pressure: number;
}

export interface RankingEntry {
  rank: number;
  org_id: string;
  name: string;
  tag: string;
  region: string;
  points: number;
  is_player_org: boolean;
}

export interface Fixture {
  tournament_id: string;
  tournament_name: string;
  opponent_name: string;
  opponent_tag: string;
  week: number;
  month: number;
  is_home?: boolean;
  round: string;
}