# lib/api-zod/src/generated/types/advanceResult.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * CS Manager game API
 * OpenAPI spec version: 0.1.0
 */
import type { GameState } from './gameState';
import type { MatchResult } from './matchResult';
import type { NewsItem } from './newsItem';

export interface AdvanceResult {
  state: GameState;
  match_results: MatchResult[];
  new_news: NewsItem[];
  week_summary: string;
}