# lib/api-zod/src/generated/types/fixture.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * CS Manager game API
 * OpenAPI spec version: 0.1.0
 */

export interface Fixture {
  tournament_id: string;
  tournament_name: string;
  opponent_name: string;
  opponent_tag: string;
  week: number;
  month: number;
  is_home?: boolean;
  round: string;
}