# lib/api-zod/src/generated/types/gameState.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * CS Manager game API
 * OpenAPI spec version: 0.1.0
 */

export interface GameState {
  has_game: boolean;
  manager_name?: string;
  week: number;
  month: number;
  year: number;
  season_phase: string;
  org_id?: string;
  org_name?: string;
  org_tag?: string;
  region?: string;
  tier?: number;
  budget: number;
  roster_size: number;
  academy_size?: number;
  form: string[];
  /** @nullable */
  world_rank?: number | null;
  chemistry?: number;
  pressure?: number;
  news_count?: number;
  week_label?: string;
}