# lib/api-zod/src/generated/types/matchResult.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * CS Manager game API
 * OpenAPI spec version: 0.1.0
 */

export interface MatchResult {
  tournament_name: string;
  team_a: string;
  team_b: string;
  score_a: number;
  score_b: number;
  winner: string;
  is_player_match: boolean;
  maps: string[];
  week: number;
}