# lib/api-zod/src/generated/types/newGameInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * CS Manager game API
 * OpenAPI spec version: 0.1.0
 */

export interface NewGameInput {
  org_id: string;
  manager_name: string;
  year?: number;
}