# lib/api-zod/src/generated/types/newsItem.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * CS Manager game API
 * OpenAPI spec version: 0.1.0
 */

export interface NewsItem {
  id: string;
  headline: string;
  category: string;
  week: number;
  month?: number;
  year?: number;
  timestamp?: string;
}