# lib/api-zod/src/generated/types/orgSummary.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * CS Manager game API
 * OpenAPI spec version: 0.1.0
 */

export interface OrgSummary {
  id: string;
  name: string;
  tag: string;
  region: string;
  tier: number;
  rating: number;
  budget: number;
  roster_size: number;
}