# lib/api-zod/src/generated/types/playerCard.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * CS Manager game API
 * OpenAPI spec version: 0.1.0
 */
import type { PlayerCardAttributes } from './playerCardAttributes';

export interface PlayerCard {
  id: string;
  name: string;
  alias: string;
  age: number;
  nationality: string;
  role: string;
  rating: number;
  hltv_rating: number;
  salary: number;
  contract_years?: number;
  form: string[];
  mental: string;
  is_academy: boolean;
  /** @nullable */
  org_id?: string | null;
  attributes?: PlayerCardAttributes;
}