# lib/api-zod/src/generated/types/rankingEntry.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * CS Manager game API
 * OpenAPI spec version: 0.1.0
 */

export interface RankingEntry {
  rank: number;
  org_id: string;
  name: string;
  tag: string;
  region: string;
  points: number;
  is_player_org: boolean;
}