# lib/api-zod/src/generated/types/rosterData.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * CS Manager game API
 * OpenAPI spec version: 0.1.0
 */
import type { PlayerCard } from './playerCard';

export interface RosterData {
  roster: PlayerCard[];
  academy: PlayerCard[];
  budget: number;
  chemistry: number;
  pressure: number;
}